pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "RENTT"
version = "0.1.0"
description = "Runtime Efficient Network to Tree Transformation (RENTT)"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Helena Monke", email = "helena.monke@ipa.fraunhofer.de"},
    {name = "Yilin Chen"},
]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "tensorflow==2.16.1",
    "torch==2.9.1",
    "keras==3.12.0",
    "pandas==2.3.3",
    "numpy==1.26.4",
    "scikit-learn==1.7.2",
    "scikit-image==0.25.2",
    "scipy==1.15.3",
    "matplotlib==3.10.7",
    "plotly==5.24.1",
    "hydra-core==1.3.2",
    "dalex==1.7.2",
    "lime==0.2.0.1",
    "sage-importance==0.0.6",
    "krippendorff==0.8.2",
    "tqdm==4.67.1",
]


[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]